Added log

+++ pkg/repositories/attendances.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import Session
 from db.postgres import engine
 from db.models import Attendance
+from logger.logger import logger
 
 
 def get_all_attendances():
@@ -23,7 +24,9 @@ def get_attendance_by_id(user_id, attendance_id):
         db_attendances = db.query(Attendance).filter(Attendance.user_id == user_id,
                                                      Attendance.id == attendance_id).first()
         if db_attendances is None:
+            logger.error(f"attendance with id {attendance_id} not found")
             return None
+
         attendance = Attendance()
         attendance.user_id = db_attendances.user_id
         attendance.lesson_id = db_attendances.lesson_id
@@ -35,31 +38,48 @@ def get_attendance_by_id(user_id, attendance_id):
 
 def create_attendance(attendance: Attendance):
     with Session(bind=engine) as db:
-        attendance_db = Attendance(user_id=attendance.user_id,
-                                   lesson_id=attendance.lesson_id,
-                                   course_id=attendance.course_id,
-                                   attended=attendance.attended,
-                                   attendance_date=attendance.attendance_date)
+        # Проверяем, есть ли уже запись с таким же user_id и lesson_id
+        existing_attendance = db.query(Attendance).filter_by(
+            user_id=attendance.user_id,
+            lesson_id=attendance.lesson_id
+        ).first()
+
+        if existing_attendance:
+            return None
+
+        # Создаём новую запись
+        attendance_db = Attendance(
+            user_id=attendance.user_id,
+            lesson_id=attendance.lesson_id,
+            course_id=attendance.course_id,
+            attended=attendance.attended,
+            attendance_date=attendance.attendance_date
+        )
+
         db.add(attendance_db)
         db.commit()
         return attendance_db.id
 
 
-def update_attendance(user_id, attendance):
+def update_attendance(attendance):
     with Session(bind=engine) as db:
-        attendance_db = db.query(Attendance).filter(Attendance.id == attendance.id,
-                                                    Attendance.user_id == user_id).first()
+        attendance_db = db.query(Attendance).filter(Attendance.id == attendance.id).first()
         attendance_db.user_id = attendance.user_id
         attendance_db.lesson_id = attendance.lesson_id
         attendance_db.course_id = attendance.course_id
         attendance_db.attended = attendance.attended
         db.commit()
+    if attendance_db is None:
+        logger.error(f"attendance with id {attendance.id} not found")
+        return None
 
 
-def delete_attendance(user_id, attendance_id):
+def delete_attendance(attendance_id):
     with Session(bind=engine) as db:
-        attendance_db = db.query(Attendance).filter(Attendance.id == attendance_id,
-                                                    Attendance.user_id == user_id).first()
+        attendance_db = db.query(Attendance).filter(Attendance.id == attendance_id).first()
+        if attendance_db is None:
+            logger.error(f"attendance with id {attendance_id} not found")
+            return None
+
         db.delete(attendance_db)
         db.commit()
-
